+++ conftest.py
@@ -1,6 +1,6 @@
 from flask import Flask
 import pytest  # type: ignore
-from main import create_app
+from app import create_app
 
 
 @pytest.fixture  # type: ignore
